+++ src/App.tsx
@@ -1,11 +1,11 @@
 import ErrorBoundary from "./components/ErrorBoundary";
 import ErrorFallback from "./components/Fallback/ErrorFallback";
-import { DataProvider } from "./context/DataContext";
+import { APIProvider } from "./context/APIContext";
 import ProductPage from "./pages/Product/ProductPage";
 
 function App() {
   return (
-    <DataProvider>
+    <APIProvider>
       <ErrorBoundary
         fallback={
           <ErrorFallback message="애플리케이션을 불러오는 중 에러가 발생했습니다." />
@@ -13,7 +13,7 @@ function App() {
       >
         <ProductPage />
       </ErrorBoundary>
-    </DataProvider>
+    </APIProvider>
   );
 }
