+++ src/pages/Product/tests/removeCartItem.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen, act, waitFor } from "@testing-library/react";
-import { DataProvider } from "@/context/DataContext";
+import { DataProvider } from "@/context/APIContext";
 import App from "@/App";
 
 const { mockProductItems } = vi.hoisted(() => {
